still adding redux, moved actions and reducers
